requirements.txt
torch
transformers
streamlit
Pillow
scikit-learn
matplotlib
